add notification system for user feedback and enhance import/export functionality
